Py Style
Apparently all of the code in this project was not conforming to the
standard python style guidelines. Especially comments. Very much the
comments. Went through all of the code files and updated them to the
correct styling as verified by the flake8. No actual code changes just
some minor refactoring and renaming when variables or class were not
named correctly.